Fix types

+++ stack/stack array/include/stk_ary.h
@@ -5,14 +5,14 @@
 typedef struct stk_ary
 {
     long top, capacity;
-    size_t *elements;
+    int *elements;
 } stk_ary_t;
 
 stk_ary_t *stk_ary_alloc(const long capacity);
 void stk_ary_dealloc(stk_ary_t **stk_ary);
-void stk_ary_push(stk_ary_t *stk_ary, const size_t element);
+void stk_ary_push(stk_ary_t *stk_ary, const int element);
 void stk_ary_pop(stk_ary_t *stk_ary);
-size_t stk_ary_peek(const stk_ary_t *stk_ary);
-void stk_ary_traverse(stk_ary_t *stk_ary, void (*fp)(size_t **element));
+int stk_ary_peek(const stk_ary_t *stk_ary);
+void stk_ary_traverse(stk_ary_t *stk_ary, void (*fp)(int **element));
 unsigned short stk_ary_is_empty(const stk_ary_t *stk_ary);
 unsigned short stk_ary_is_full(const stk_ary_t *stk_ary);
